Get user properties directly for author pages
Calling `WP_User::get()` to retrieve user nicename and display name for
author pages seems a bit non-standard, as that's just a wrapper around
the magic getter anyways. I've found this to cause fatals when
integrating the with plugins like Bylines which override
`get_queried_object()` on author pages but don't provide the `get()`
method on the object used in place of a user object.

Accessing the nicename and display name properties directly (and letting
the object's getter methods handle the implementation, as is being done
with other objects in this function) seems like a more standard way of
doing this.